faiss-sphere-new/requirements.txt
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
faiss-cpu>=1.7.3
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0
pytest>=7.0.0

# Optional: for real Wikipedia data
# datasets>=2.0.0
# transformers>=4.20.0
# torch>=1.10.0
